+++ packages/spec-test-runner/test/utils/specTestCases.ts
@@ -1,5 +1,5 @@
-import {join} from "path";
+import path from "path";
 
 export const SPEC_TEST_REPO_URL = "https://github.com/ethereum/eth2.0-spec-tests";
 export const SPEC_TEST_VERSION = "v1.1.0-alpha.5";
-export const SPEC_TEST_LOCATION = join(__dirname, "../../spec-tests");
+export const SPEC_TEST_LOCATION = path.join(__dirname, "../../spec-tests");
