+++ packages/validator/.mocharc.yaml
@@ -1,3 +1,3 @@
 colors: true
 require: ts-node/register
-file: ./test/unit/setup.ts
+file: ./test/setup.ts
