+++ packages/validator/package.json
@@ -29,7 +29,8 @@
     "pretest": "yarn run check-types",
     "test:unit": "nyc --cache-dir .nyc_output/.cache -e .ts mocha 'test/unit/**/*.test.ts'",
     "test": "yarn test:unit",
-    "test:e2e": "mocha --timeout 5000 'test/e2e/**/*.test.ts'",
+    "test:e2e": "yarn run download-test-cases && mocha 'test/e2e/**/*.test.ts'",
+    "download-test-cases": "node -r ts-node/register test/e2e/slashing-protection-interchange-tests/downloadTests.ts",
     "coverage": "codecov -F lodestar-validator",
     "benchmark": "ts-node test/benchmarks"
   },
