+++ packages/validator/src/slashingProtection/interchange/parseInterchange.ts
@@ -14,31 +14,7 @@ export function parseInterchange(
   const format = (interchange as IInterchangeCompleteV4)?.metadata?.interchange_format;
   const version = interchange?.metadata?.interchange_format_version;
 
-  // version < v5.0.0 (older version)
-  if (format) {
-    switch (format) {
-      case "complete":
-        switch (version) {
-          case "4": {
-            const interchangeLodestar = parseInterchangeCompleteV4(config, interchange as IInterchangeCompleteV4);
-            if (!isEqualRoot(config, interchangeLodestar.genesisValidatorsRoot, expectedGenesisValidatorsRoot)) {
-              throw new InterchangeError({
-                code: InterchangeErrorErrorCode.GENESIS_VALIDATOR_MISMATCH,
-                root: interchangeLodestar.genesisValidatorsRoot,
-                extectedRoot: expectedGenesisValidatorsRoot,
-              });
-            }
-            return interchangeLodestar;
-          }
-
-          default:
-            throw new InterchangeError({code: InterchangeErrorErrorCode.UNSUPPORTED_VERSION, version});
-        }
-
-      default:
-        throw new InterchangeError({code: InterchangeErrorErrorCode.UNSUPPORTED_FORMAT, format: format});
-    }
-  } else {
+  if (!format) {
     // version >= v5.0.0
     switch (version) {
       case "5": {
@@ -56,5 +32,29 @@ export function parseInterchange(
       default:
         throw new InterchangeError({code: InterchangeErrorErrorCode.UNSUPPORTED_VERSION, version});
     }
+  }
+
+  // version < v5.0.0 (older version)
+  switch (format) {
+    case "complete":
+      switch (version) {
+        case "4": {
+          const interchangeLodestar = parseInterchangeCompleteV4(config, interchange as IInterchangeCompleteV4);
+          if (!isEqualRoot(config, interchangeLodestar.genesisValidatorsRoot, expectedGenesisValidatorsRoot)) {
+            throw new InterchangeError({
+              code: InterchangeErrorErrorCode.GENESIS_VALIDATOR_MISMATCH,
+              root: interchangeLodestar.genesisValidatorsRoot,
+              extectedRoot: expectedGenesisValidatorsRoot,
+            });
+          }
+          return interchangeLodestar;
+        }
+
+        default:
+          throw new InterchangeError({code: InterchangeErrorErrorCode.UNSUPPORTED_VERSION, version});
+      }
+
+    default:
+      throw new InterchangeError({code: InterchangeErrorErrorCode.UNSUPPORTED_FORMAT, format: format});
   }
 }
